requirements.txt
Click
Flask
Jinja2
MarkupSafe
TwitterSearch
Werkzeug
beautifulsoup4
bs4
certifi
chardet
future
idna
itsdangerous
nltk
numpy
oauthlib
pip
profanity
profanity-check
python-twitter
requests
requests-oauthlib
scikit-learn
scipy
setuptools
six
soupsieve
urllib3